Optimise handling of missing data with `ets:lookup_element/3`
A common pattern is to write code of the form
`try ets:lookup_element(A,B,C) of ... catch ... end` to handle missing
data in an ETS table, since `ets:lookup_element/3` would throw an
error in that case. More recently `ets:lookup_element/4` has been
added, which introduces a 4th parameter to use as a default parameter
in the case that there is no data in the ETS table corresponding to
the given table and key. The new way is faster and more ergonomic
(see GitHub #6234), so we migrate to it here.