Implement DNS LOC record (RFC 1876)
The record data Erlang representation is chosen to
be similar to the Master File Format described in RFC section 3,
i.e latitude and longitude in degrees, and altitude, size and precision
in meters; all as floating point values.

Encoding also supports integers with latitude and longitude
in thounsands of an arc second, and alitude, size and precision
in centimeters, as the RDATA format in RFC section 2.

Encoding of size and precision encodes the smallest encoding
value not less than the input value; e.g 100 encodes as 100,
4711 encodes as 5000, and 90001 encodes as 100000.