compiler: Be stricter when tracking values in private append
Tracking values to patch into writable binaries in the private append
pass doesn't make any attempt to use type information to exclude
execution paths not relevant when tracking an appendable binary. In
order to avoid later failed attempts to rewrite non-bitstring literals
into bs_init_writable-created values during the rewrite phase, we
change the tracking phase to preemptively abort tracking as soon as we
detect that the element we are tracking can't apply to the literal.

Closes #6999