erts: Require modules to be made writable before modifying code
Rather than allowing modules to be writable all the time, either
through RWX pages or a dual-mapped RW page, support making them
writable only in the brief window that they are actually being
modified.

To avoid introducing subtle bugs this late in the release cycle,
we'll settle for adding assertions that modifications are only
made while a module is marked writable. We won't change any page
permissions just yet.

The one exception to this are the per-thread permissions on MacOS,
where a thread can temporarily trade permission to execute MAP_JIT
pages for being able to write to them. This permission is added
when making a module writable and removed when we're done. This
fixes the long-standing issue GH-4911 on Apple Silicon (Intel macs
are still affected since we don't want to give up W^X).