compiler: Avoid crash during value tracking in private append
In the private append pass we track values which could be bitstrings
or a term containing a bitstring to find the operation where the
bitstring is constructed. For the purpose of this description, we call
such a value a potential bitstring.

As tracking potential bitstrings is done without correlating type
checks and flow-control, we can sometimes encounter values which are
not potential bitstrings. In 0bac830c937c7c39b6aae37a73bd135841dde969
the tracking logic was refined to abort as soon as we detected that a
returned literal could not be on an execution path returning a
potential bitstring. This commit extends the tracking logic to abort
tracking when a potential bitstring is constructed by an operation
which cannot produce a bitstring or a term containing a bitstring.

Closes #7011