compiler: Remove dead code in private append
When 0bac830c937c7c39b6aae37a73bd135841dde969 was created, the element
forms to be considered were taken from patch_literal_term/3. As the
tuple_elements-form is only created after the tracking phase, it will
never be encountered in get_results/5. Likewise the Element variable
is never a list, as it is generated by patch_appends_ret/3