[kernel] Use async_dist in net_kernel instead of nosuspend tricks
net_kernel used to avoid getting blocked when sending replies by using
send with nosuspend and spawn a process doing the reply if the send operation
would have suspended. Using async_dist we prevent the blocking, keep the
signal order, avoid one extra copying of the reply, avoid the overhead
of creating a process, and avoid the extra memory consumption due to the
extra process.

net_kernel also used noconnect in replies since it would deadlock when
connection setup was syncronous which is no longer needed.

Due to the two above options, it also did not use the gen_server:reply/2
API which caused it not to reply using aliases. Now all replies are made
using gen_server:reply/2.