ssl: Enhance warning alert handling
If a user cancel alert with level warning is received during handshake make
it be handled the same regardless of TLS version. If it is received in the state connection
in TLS-1.3 regard it as an error as it is inappropriate.

In TLS-1.3 all error alerts are considered FATAL regardless of legacy
alert type. But make sure legacy type is printed in logs to not confuse users
that are expecting the same legacy type as sent by peer.