gh: Add a label which forces all tests to be run
Adds a check in CI that looks for the `full-build-and-check`
label, and if it is present, overrides the checking of which
apps have changed, and tests everything.

This is useful for changes which can have consequences beyond the
modules which were directly modified, giving confidence without
needing to wait for the daily tests.