internal_export feature
Feature `internal_export` is added as experimental.
Simply, feature looks at functions tagged with "internal_export" attribute and adds them to the list of exports (while keepining "internal_export" attribute as it is).
This is just to avoid code duplication.
This way it will be easier to support internal_exports as a top level beam chunk if some day it gets implemented.