fix(supervisor): fix return value when process is already registered
This fixes a bug where the return type for `supervisor:start_child/2`
was not being respected.

If a supervisor tries to start a child who tries to register the same
local name that a running process has already registered, the return
value was of the form `{error, {{already_started, pid()}, child_spec()}}`
rather than `{error, {already_started, pid()}}`.