[erts|esock] Fixed reading with zero (0) timeout
Fixing handling reading (recv, recvfrom and recvmsg) with
timeout = 0 (zero).
This feature was used by the api_opt_ip_recverr_udp4 test
case. This test case is unlikely to work on Windows, but
have been updated so that it *can* work (but will most
likely result in a skip).

OTP-18029