[erts] Fix auto connect from net_kernel
This bug was harmless. Prior to ff7daaca438743cc14b79ed73c2fd6a1e7b46307
(on master prior to OTP-26.0-rc3) net_kernel never triggered auto connect
itself. After that commit, when net_kernel triggered auto connect, the
runtime system would try-lock the main-lock on net_kernel while executing
net_kernel when sending a message to net_kernel. An assertion would trigger,
but in the optimized runtime this would just cause the try-lock to fail
and the message to unnecessarily to be placed in a heap-fragment.