kernel: make printing strings with ANSI SGR escapes faster
Printing a long string with ANSI escapes is very slow on Erlang/OTP 26 RC2
which is caused by calling re:run/3 nifs a bunch of times.
Compiling the regex did not make it fast enough, so SGR
(Select Graphic Rendition) escapes are dealt separately with an erlang
function which is what most people use escapes for.
See #7040