Minor refactor: use erts_cleanup_offheap_list in erl_db_util.c
There are multiple places in erl_db_util.c where a temporary ErlOffHeap
is created just to call erts_cleanup_offheap(). This function then
extracts the first field, and passes it to erts_cleanup_offheap_list().
Calling that function directly is simpler and removes the need to create
the temporary object.