inet_res: fix edns0 unit test
Most calls to inet_res:resolve are missing 'nameservers' which
means the request goes to your local resolver which may route
single labels to LLMNR (eg. systemd-resolved). Also as Ericsson
did not buy the TLD 'otptest' these are not publically usable.

Also, inet_res:getbyname() does not accept options and we are
not going to hijack the inet_res main resolver, so remove it.