Improve documentation for gen_tcp:recv/3
The first part of this patch only highlights the word "passive" in the
documentation for gen_tcp:recv/3, to emphasize that this function only
works in passive mode.

The second part adds a note saying that calling gen_tcp:recv/3 on
an active socket returns {error, einval}.

The last part of the patch adds a paragraph to the documentation for
gen_tcp:recv/3 mentioning what happens when multiple processes call
recv/3 on the same socket. This was previously not documented.