[erts|esock] Add function for handling 'more_data' error
When on Windows reading with recvfrom, if the buffer is
too small for the data available, the result is an error:
more_data. Add a function for hanling this special error case.
Also "fixed" buffer size calculation for recvfrom read buffer
(ensure it is a minimum of 32 KB). *Workaround*.

OTP-18586