diameter:  remove optional call to deprecated fn and update appup
`diameter` should not call or rely on the deprecated function
`dbg:stop_clear/0`.

The main reason is that `dbg:stop_clear/0` is simply an alias to
`dbg:stop/0`. Thus, `dbg:stop_clear/0` was marked as deprecated in a
previous pull request GH-6903 (commit
`861be72c5a1261ee1694ee468540256f6db11e87`), and these are the remains
that I forgot to update.