compiler: Use types to avoid false aliasing with get_tuple_element
Use type information to avoid false aliasing when extracting elements
from a tuple using the `get_tuple_element` instruction. Extracting an
element known to be of a plain type always results in an unique and
does not change the alias status of the tuple.

Also add test case that will pass when type information is fully
exploited.