Merely warn for unbound variables in union types
In #6864, it became an error to have unbound variables in a union.
Before that, unbound variables in unions were silently ignored.

For compatibility with code that used to compile before Erlang/OTP 26,
change the error to a warning. The warning can be disabled with the
`nowarn_singleton_typevar` option.

It is expected that unbound variables in unions will again become an
error in Erlang/OTP 27.

Closes #7116