ssl: Add signature_algs/2
To make it easier to configure signature algorithms with algorithms that
are moved from the default add the API function signature_algs/2 that
lists possible values.

Also adjust signature algorithms according to TLS-1.3 RFC. That is TLS-1.3
legacy algorithms/schemes names should be used insted of TLS-1.2 algorithms names
to be supported in TLS-1.2-TLS-1.3 combination. sha224 is also considered legacy
and should not be part of the defaults.