[erts] Improved signal handling
Keep track of the amount of message signals between non-message signals
and update message queue length when messages are moved into the message
queue.

This simplifies the implementation of process_info(_, message_queue_len)
a lot, and also makes it possible to make an efficient implementation of
message queue length monitoring.