erts: Fix optimization bug in maps:put
maps:put with existing key and identical value was not optimized
as a no-op correctly in collision node.

Bug was harmeless though as CDR(ptr) is the first element of the
collision tuple which is an internal key-value cons that can never
be equal to 'value'. That is, the optimization cannot kick in
and yield incorrect result.