compiler: Allow ssa_opt to run on modules with NIFs
This patch allows the beam_ssa_opt-pass to run on modules containing
NIFs.

NIFs are problematic for the SSA optimization passes as when a loaded
NIF replaces a function, essentially all bets are off as callers of
the NIF cannot make any assumptions on the result of calling the NIF.
Traditionally, the SSA optimization passes have completely ignored
modules containing a call to erlang:load_nif/2. A safe way to handle
NIFs, but still allow optimization of modules containing NIFs is to
make calls to NIFs look like external calls to the optimization
passes.

The function isolate_nifs/1 transforms the input #b_module{} by
rewriting all calls to NIFs in the module to to calls to external
functions with the same names in another module. As this also removes
all callers to the NIF, all non-exported NIFs are forcibly exported,
to avoid them being removed as dead code. As all passes know how
handle external calls, this allows for safe optimization of the
module. That a NIF function can contain BEAM code which calls other
functions in the module is not a problem, at worst it leads to missed
optimizations. When all sub-passes of beam_ssa_opt have been executed
restore_nifs/2 undoes the module transforms done by isolate_nifs/1.

To avoid extra book-keeping to keep track of rewritten calls for use
by restore_nifs/2, the module to which the calls are redirected is
given a name, '\nnifs', which cannot be created by the user.

As it can be assumed that the existing beam_ssa_opt sub-passes handle
external calls properly, testing of this patch is limited to checking:
that the result of a NIF is not constant propagated; that the presence
of a call to erlang:load_nif/2 does not disable optimization of
functions not calling a NIF; and that non-exported NIFs, stay
non-exported (that exported NIFs, stay exported is already covered by
the nif_SUITE tests).