compiler: bif:binary_part/{2,3} should alias its arguments
The `bif:binary_part/{2,3}` is a guard bif which creates a
sub-binary (containing a reference to its argument) from its argument,
thus aliasing both. Currently this does not produce any errors as the
BEAM will only ever append to binaries, but if we ever attempt to do
destructive updates of binaries this will give us grief, so we'd
better correct it.