compiler: SSA-checker: Allow pattern variables in annotation maps
This patch adds support for matching pattern variables when a map is
used as a value in an annotation. An interesting detail is that, as
variables are now allowed in annotations, we want to check the
annotation after the operation, as it may bind variables.

This change lacks a test case in
beam_ssa_check_SUITE_data/sanity_checks.erl as the compiler does not
yet produce any annotations with maps containing SSA-variables. It
will be added in a future patch.