Fix is_record/3 for unreasonable tuple sizes
Calling `is_record/3` with a zero tuple size would crash the compiler.
For example:

    foo(A) -> is_record(A, a, 0).

Using the same kind of expression in a guard would generate incorrect
code that could succeed. For example, given:

    bar(A) when is_record(A, a, 0) -> ok.

the call `bar({a})` would succeed.

The compiler would run out of memory or never finish compiling a call
to `is_record/3` with a huge tuple size. Example:

    foobar(A) when is_record(A, a, 10000000) -> A.

Closes #7298
Closes #7317