[erts|esock] Add support for UDS on Windows
Add support for Unix Domain Sockets on Windows.
Only minor tweaks needed. One part was that the
'configure' tests for this does not seem to "work"
on Windows, so we add our own much simpler test.
Also, extend the allowed domains for accept (previously
only handled inet and inet6).

OTP-18611