erts: Remove buggy optimization in fun purging
Calling a fun defined by a module that had been purged after
loading a different version of the same module (and therefore did
not inherit the old fun entries) could cause the emulator to
crash.

We could retain the optimization by comparing the MD5 of the old
and current instances, but it didn't give much so it's better to
just get rid of it.

Fixes #7288