Compile escripts by default
The current default when running an escript is to interpret it using
the `erl_eval` module. In the ancient times when escript was invented
that made sense. Interpretation would be slower and some language
constructs would not work, but execution could start immediately
without waiting for the compiler to finish.

Nowadays with faster computers it makes less sense.

Therefore, in Erlang/OTP 27, change the default to compile the
escript. That will require the escript to be executed by an Erlang
system that includes the Erlang compiler.