Extend PropEr with atomlimit-safe generator variants
The atom generator of PropEr generates atoms from random strings
and its use, explicitly or implicitly, is prone to exhausting
the atom limit.

This commit adds variants to the atom generator which pick from
the existing atoms and do not generate any new ones. It also provides
variants of the any, list, map, term and tuple generators which implicitly
use the atom generator.

This extension is intended for internal use in property-based tests
in OTP. It will only be enabled when PropEr is detected as property
testing tool.

Co-authored-by: Jan Uhlig <juhlig@hnc-agency.org>
Co-authored-by: Björn Gustavsson <bjorn@erlang.org>