compiler: Fix bug in alias analysis when extracting from a map
When a value is extracted from a map and the map dies, the extracted
value should inherit the status of the map. If the map does not die,
both the map and the extracted value should be aliased.

This patch corrects an error in the alias analysis pass which failed
to propagate a status of 'aliased' from the map to the extracted
value. It also unifies the handling of the BIF `map_get` and the
instruction `get_map_element`. The patch also adds test-cases to check
that both the BIF and the instruction behaves the same.