compiler: Fix bug in alias analysis when extracting from a pair
When the BIFs `hd/1` and `tl/1` are used to extract an element from a
pair, the results are wrong and do not match the result of using the
`get_hd` and `get_tl` instructions. This patch corrects the error by
using the same code as for instructions to handle the BIFs.