compiler: Account for extracted elements when doing alias analysis
When elements are extracted from pairs and tuples, the extracted
element forms an alias to the element in the original aggregate. But
as long as the reference to the original is only used to extract other
elements and the original reference is not use for anything else, it
does not lead to any aliasing that is an issue for destructive updates.

This patch enhances the alias analysis to account for extracted
elements. This is done by extending `aa_get_status/2` to consider a
reference from which elements has been extracted as aliased. Likewise
`aa_set_status/3` now updates the status of all extracted elements if
the original reference becomes aliased. Correspondingly
`aa_tuple_extraction/2` and `aa_pair_extraction/4` are modified to
disregard extracted elements as long as the element being extracted
has not already been extracted.

As this enhancement will let the private-append pass explore new
execution paths which were ignored previously due to aliasing, the
value tracking is updated to handle mismatched element types.