emulator: Remove vestiges of special handling of x(0)
In the BEAM emulator, X register 0 (`x(0)`) used to be specially
handled by the macro `r(0)` so that it could be placed in a CPU
register.

That made sense for platforms with many registers such as Sparc and
PowerPC. For the register-starved x86_64 platform it is not possible
to allocate a CPU register for `x(0)`. Therefore, in
1f996cc46a8c93 (part of Erlang/OTP 19), the special handling of `x(0)`
was removed. Removing that special case simplified the code in many
places.

Get rid of the vestiges of the special handling of `x(0)`. While at
it, also re-introduce the `REG_xregs` attribute for the variable
holding the pointer to the X registers.