tty: Fix blocking ReadFile call on Windows
When calling ReadFile using blocking semantics on a
terminal, the call would block until a line has been
entered on the console. This makes it impossible
to interact with the process doing the read (to change
encoding for example).

So we change to using non-blocking (OVERLAPPED) ReadFile
when reading from a Console when there is no shell started.

Fixes #7261